+++ src/Farmer/Builders/Builders.Dns.fs
@@ -1,6 +1,8 @@
 [<AutoOpen>]
 module Farmer.Builders.Dns
 
+open System
+
 open Farmer
 open Farmer.Dns
 open Farmer.Arm.Dns
@@ -27,10 +29,21 @@ type NsRecordProperties =  { Name: ResourceName; NsdNames : string list; TTL: in
 type PtrRecordProperties =  { Name: ResourceName; PtrdNames : string list; TTL: int option; }
 type TxtRecordProperties =  { Name: ResourceName; TxtValues : string list; TTL: int option; }
 type MxRecordProperties =  { Name: ResourceName; MxValues : {| Preference : int; Exchange : string |} list; TTL: int option; }
+type SrvRecordProperties =  { Name: ResourceName; SrvValues : SrvRecord list; TTL: int option; }
+type SoaRecordProperties =
+    { Name: ResourceName
+      Host : string option
+      Email : string option
+      SerialNumber : int64 option
+      RefreshTime : int64 option
+      RetryTime : int64 option
+      ExpireTime : int64 option
+      MinimumTTL : int64 option
+      TTL: int option }
 
 type DnsCNameRecordBuilder() =
-    member __.Yield _ = { CNameRecordProperties.CName = None; Name = ResourceName.Empty; TTL = None; TargetResource = None }
-    member __.Run(state : CNameRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, CName(state.TargetResource, state.CName))
+    member _.Yield _ = { CNameRecordProperties.CName = None; Name = ResourceName.Empty; TTL = None; TargetResource = None }
+    member _.Run(state : CNameRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, CName(state.TargetResource, state.CName))
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
@@ -50,13 +63,13 @@ type DnsCNameRecordBuilder() =
     member _.RecordTargetResource(state:CNameRecordProperties, targetResource) = { state with TargetResource = Some targetResource }
 
 type DnsARecordBuilder() =
-    member __.Yield _ = { ARecordProperties.Ipv4Addresses = []; Name = ResourceName "@"; TTL = None; TargetResource = None }
-    member __.Run(state : ARecordProperties)  = DnsZoneRecordConfig.Create(state.Name, state.TTL, A(state.TargetResource, state.Ipv4Addresses))
+    member _.Yield _ = { ARecordProperties.Ipv4Addresses = []; Name = ResourceName "@"; TTL = None; TargetResource = None }
+    member _.Run(state : ARecordProperties)  = DnsZoneRecordConfig.Create(state.Name, state.TTL, A(state.TargetResource, state.Ipv4Addresses))
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:ARecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:ARecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Sets the ipv4 address.
     [<CustomOperation "add_ipv4_addresses">]
@@ -71,13 +84,13 @@ type DnsARecordBuilder() =
     member _.RecordTargetResource(state:ARecordProperties, targetResource) = { state with TargetResource = Some targetResource }
 
 type DnsAaaaRecordBuilder() =
-    member __.Yield _ = { AaaaRecordProperties.Ipv6Addresses = []; Name = ResourceName "@"; TTL = None; TargetResource = None }
-    member __.Run(state : AaaaRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, AAAA(state.TargetResource, state.Ipv6Addresses))
+    member _.Yield _ = { AaaaRecordProperties.Ipv6Addresses = []; Name = ResourceName "@"; TTL = None; TargetResource = None }
+    member _.Run(state : AaaaRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, AAAA(state.TargetResource, state.Ipv6Addresses))
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:AaaaRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:AaaaRecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Sets the ipv6 address.
     [<CustomOperation "add_ipv6_addresses">]
@@ -92,13 +105,13 @@ type DnsAaaaRecordBuilder() =
     member _.RecordTargetResource(state:AaaaRecordProperties, targetResource) = { state with TargetResource = Some targetResource }
 
 type DnsNsRecordBuilder() =
-    member __.Yield _ = { NsRecordProperties.NsdNames = []; Name = ResourceName "@"; TTL = None; }
-    member __.Run(state : NsRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, NS state.NsdNames)
+    member _.Yield _ = { NsRecordProperties.NsdNames = []; Name = ResourceName "@"; TTL = None; }
+    member _.Run(state : NsRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, NS state.NsdNames)
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:NsRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:NsRecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Add NSD names
     [<CustomOperation "add_nsd_names">]
@@ -114,8 +127,8 @@ type DnsPtrRecordBuilder() =
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:PtrRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:PtrRecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Add PTR names
     [<CustomOperation "add_ptrd_names">]
@@ -126,13 +139,13 @@ type DnsPtrRecordBuilder() =
     member _.RecordTTL(state:PtrRecordProperties, ttl) = { state with TTL = Some ttl }
 
 type DnsTxtRecordBuilder() =
-    member __.Yield _ = { TxtRecordProperties.Name = ResourceName "@"; TxtValues = []; TTL = None; }
-    member __.Run(state : TxtRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, TXT state.TxtValues)
+    member _.Yield _ = { TxtRecordProperties.Name = ResourceName "@"; TxtValues = []; TTL = None; }
+    member _.Run(state : TxtRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, TXT state.TxtValues)
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:TxtRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:TxtRecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Add TXT values
     [<CustomOperation "add_values">]
@@ -143,13 +156,13 @@ type DnsTxtRecordBuilder() =
     member _.RecordTTL(state:TxtRecordProperties, ttl) = { state with TTL = Some ttl }
 
 type DnsMxRecordBuilder() =
-    member __.Yield _ = { MxRecordProperties.Name = ResourceName "@"; MxValues = []; TTL = None; }
-    member __.Run(state : MxRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, MX state.MxValues)
+    member _.Yield _ = { MxRecordProperties.Name = ResourceName "@"; MxValues = []; TTL = None; }
+    member _.Run(state : MxRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, MX state.MxValues)
 
     /// Sets the name of the record set.
     [<CustomOperation "name">]
-    member _.RecordName(state:CNameRecordProperties, name) = { state with Name = name }
-    member this.RecordName(state:CNameRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+    member _.RecordName(state:MxRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:MxRecordProperties, name:string) = this.RecordName(state, ResourceName name)
 
     /// Add MX records.
     [<CustomOperation "add_values">]
@@ -160,6 +173,88 @@ type DnsMxRecordBuilder() =
     /// Sets the TTL of the record.
     [<CustomOperation "ttl">]
     member _.RecordTTL(state:MxRecordProperties, ttl) = { state with TTL = Some ttl }
+
+type DnsSrvRecordBuilder() =
+    member _.Yield _ = { SrvRecordProperties.Name = ResourceName "@"; SrvValues = []; TTL = None; }
+    member _.Run(state : SrvRecordProperties) = DnsZoneRecordConfig.Create(state.Name, state.TTL, SRV state.SrvValues)
+
+    /// Sets the name of the record set.
+    [<CustomOperation "name">]
+    member _.RecordName(state:SrvRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:SrvRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+
+    /// Add SRV records.
+    [<CustomOperation "add_values">]
+    member _.RecordValue(state:SrvRecordProperties, srvValues : SrvRecord list) =
+        { state with SrvValues = state.SrvValues @ srvValues }
+
+    /// Sets the TTL of the record.
+    [<CustomOperation "ttl">]
+    member _.RecordTTL(state:SrvRecordProperties, ttl) = { state with TTL = Some ttl }
+
+type DnsSoaRecordBuilder() =
+    member _.Yield _ =
+        { SoaRecordProperties.Name = ResourceName "@"
+          Host = None
+          Email = None
+          SerialNumber = None
+          RefreshTime = None
+          RetryTime = None
+          ExpireTime = None
+          MinimumTTL = None
+          TTL = None }
+
+    member _.Run(state : SoaRecordProperties) =
+        let value =
+            { Host = state.Host
+              Email = state.Email
+              SerialNumber = state.SerialNumber
+              RefreshTime = state.RefreshTime
+              RetryTime = state.RetryTime
+              ExpireTime = state.ExpireTime
+              MinimumTTL = state.MinimumTTL }
+        DnsZoneRecordConfig.Create(state.Name, state.TTL, SOA value)
+
+    /// Sets the name of the record set.
+    [<CustomOperation "name">]
+    member _.RecordName(state:SoaRecordProperties, name) = { state with Name = name }
+    member this.RecordName(state:SoaRecordProperties, name:string) = this.RecordName(state, ResourceName name)
+
+    /// Sets the name of the record set.
+    [<CustomOperation "host">]
+    member _.RecordHost(state:SoaRecordProperties, host : string) = { state with Host = Some host }
+
+    /// Sets the email for this SOA record (required).
+    [<CustomOperation "email">]
+    member _.RecordEmail(state:SoaRecordProperties, email : string) = { state with Email = Some email }
+
+    /// Sets the expire time for this SOA record in seconds.
+    /// Defaults to 2419200 (28 days).
+    [<CustomOperation "expire_time">]
+    member _.RecordExpireTime(state:SoaRecordProperties, expireTime : int64) = { state with ExpireTime = Some expireTime }
+
+    /// Sets the minimum time to live for this SOA record in seconds.
+    /// Defaults to 300.
+    [<CustomOperation "minimum_ttl">]
+    member _.RecordMinimumTTL(state:SoaRecordProperties, minTTL : int64) = { state with MinimumTTL = Some minTTL }
+
+    /// Sets the refresh time for this SOA record in seconds.
+    /// Defaults to 3600 (1 hour)
+    [<CustomOperation "refresh_time">]
+    member _.RecordRefreshTime(state:SoaRecordProperties, refreshTime : int64) = { state with RefreshTime = Some refreshTime }
+
+    /// Sets the retry time for this SOA record in seconds.
+    /// Defaults to 300 seconds.
+    [<CustomOperation "retry_time">]
+    member _.RetryTime(state:SoaRecordProperties, retryTime : int64) = { state with RetryTime = Some retryTime }
+
+    /// Sets the serial number for this SOA record (required).
+    [<CustomOperation "serial_number">]
+    member _.RecordSerialNumber(state:SoaRecordProperties, serialNo : int64) = { state with SerialNumber = Some serialNo }
+
+    /// Sets the TTL of the record.
+    [<CustomOperation "ttl">]
+    member _.RecordTTL(state:SoaRecordProperties, ttl) = { state with TTL = Some ttl }
 
 type DnsZoneConfig =
     { Name : ResourceName
@@ -210,3 +305,5 @@ let nsRecord = DnsNsRecordBuilder()
 let ptrRecord = DnsPtrRecordBuilder()
 let txtRecord = DnsTxtRecordBuilder()
 let mxRecord = DnsMxRecordBuilder()
+let srvRecord = DnsSrvRecordBuilder()
+let soaRecord = DnsSoaRecordBuilder()